agent/requirements.txt
watchdog>=4.0.0
customtkinter>=5.2.2
pystray>=0.19.5
Pillow>=10.0.0
